some refactoring
add .gitignore
add idgenerator interface
