+++ src/main/java/com/justcodeit/moyeo/study/common/PostIdGenerator.java
@@ -1,0 +1,7 @@
+package com.justcodeit.moyeo.study.common;
+
+public interface PostIdGenerator {
+
+  String postId();
+
+}
