+++ src/main/java/com/justcodeit/moyeo/study/common/RandomIdUtil.java
@@ -3,12 +3,14 @@ package com.justcodeit.moyeo.study.common;
 import java.time.Instant;
 import java.util.concurrent.atomic.AtomicInteger;
 
-public class RandomIdUtil {
+public class RandomIdUtil implements UserIdGenerator, PostIdGenerator {
+
   private static final int counterMax = 256 * 256;
   private static final AtomicInteger intVal = new AtomicInteger(0);
 
   private static String generate() {
-    final long uid = Instant.now().toEpochMilli() * counterMax + intVal.accumulateAndGet(1, (index, inc) -> (index + inc) % counterMax);
+    final long uid = Instant.now().toEpochMilli() * counterMax + intVal.accumulateAndGet(1,
+        (index, inc) -> (index + inc) % counterMax);
 
     return Long.toHexString(uid);
   }
@@ -17,4 +19,13 @@ public class RandomIdUtil {
     return generate();
   }
 
+  @Override
+  public String userId() {
+    return "U" + next();
+  }
+
+  @Override
+  public String postId() {
+    return "P" + next();
+  }
 }
