+++ src/main/java/com/justcodeit/moyeo/study/common/UserIdGenerator.java
@@ -1,0 +1,7 @@
+package com.justcodeit.moyeo.study.common;
+
+public interface UserIdGenerator {
+
+  String userId();
+
+}
