--- src/main/java/com/justcodeit/moyeo/study/interfaces/dto/EmptyObjRes.java
@@ -1,15 +1,0 @@
-package com.justcodeit.moyeo.study.interfaces.dto;
-
-import java.util.HashMap;
-import java.util.Map;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-
-@Data
-@EqualsAndHashCode(callSuper = false)
-public class EmptyObjRes extends SuccessRes<Map<String, Object>> {
-
-  public EmptyObjRes() {
-    super(new HashMap<>());
-  }
-}
