+++ src/main/java/com/justcodeit/moyeo/study/interfaces/resource/SkillController.java
@@ -1,7 +1,6 @@
 package com.justcodeit.moyeo.study.interfaces.resource;
 
 import com.justcodeit.moyeo.study.application.skill.SkillService;
-import com.justcodeit.moyeo.study.interfaces.dto.SuccessRes;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -15,7 +14,7 @@ public class SkillController {
     private final SkillService skillService;
 
     @PostMapping
-    public ResponseEntity<SuccessRes<Boolean>> uploadZip() {
-        return ResponseEntity.ok(new SuccessRes<>(skillService.skillSaveAllFromS3()));
+    public ResponseEntity<Boolean> uploadZip() {
+        return ResponseEntity.ok(skillService.skillSaveAllFromS3());
     }
 }
