+++ src/main/java/com/justcodeit/moyeo/study/model/skill/SkillCategoryConverter.java
@@ -1,6 +1,6 @@
 package com.justcodeit.moyeo.study.model.skill;
 
-import com.justcodeit.moyeo.study.persistence.SkillCategory;
+import com.justcodeit.moyeo.study.model.type.SkillCategory;
 
 public class SkillCategoryConverter {
     public static SkillCategory getForFolderName(String skillEngWord) {
