+++ src/main/java/com/justcodeit/moyeo/study/persistence/Skill.java
@@ -1,9 +1,6 @@
 package com.justcodeit.moyeo.study.persistence;
 
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-import lombok.RequiredArgsConstructor;
-
+import com.justcodeit.moyeo.study.model.type.SkillCategory;
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
@@ -11,6 +8,8 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
 
 @Entity
 @Table(name = "skill")
